srgan/requirements.txt
numpy >= 1.17.0
scikit-image >= 0.18.1
pillow >= 8.2.0